Fix tool setting ui.
