+++ src/ai_code_assistant/gui/gui_app.py
@@ -11,6 +11,7 @@ from dotenv import load_dotenv
 
 from ai_code_assistant.gui.view.chat_page import chat_ui
 from ai_code_assistant.gui.view.chat_state import ChatUiMessage
+from ai_code_assistant.gui.view.tool_settings.tool_widget import tool_ui
 from ai_code_assistant.gui.view_model.ai_assistant_state import AiAssistantViewModel
 from ai_code_assistant.utils.logger import setup_logger
 
@@ -22,16 +23,25 @@ assistant_model = AiAssistantViewModel()
 
 
 @me.page(
-    path="/ai_code_assistant",
+    path="/",
     title="AI Code Assistant",
     security_policy=me.SecurityPolicy(dangerously_disable_trusted_types=True),
 )  # type: ignore[misc]
-def app() -> None:
+def main_page() -> None:
     chat_ui(
         transform,
         title="AI Code Assistant",
         bot_user="AI Assistant",
     )
+
+
+@me.page(
+    path="/tool_settings",
+    title="Tool Settings",
+    security_policy=me.SecurityPolicy(dangerously_disable_trusted_types=True),
+)  # type: ignore[misc]
+def tool_settings_page() -> None:
+    tool_ui()
 
 
 def transform(sentence: str, system_prompt: str, _: list[ChatUiMessage]) -> Generator[str, None, None]:
