+++ src/ai_code_assistant/gui/view/tool_settings/tool_state.py
@@ -2,13 +2,11 @@
 #
 #  This software is released under the MIT License.
 #  http://opensource.org/licenses/mit-license.php
-
 import mesop as me
 
 
 @me.stateclass
 class ToolState:
-    modal_shown: bool = False
-    rewrite: str = ""
+    clone_url: str = ""
     preview_original: str = ""
     preview_rewrite: str = ""
